// src/Veilora.Api/ClientApp/src/types/character.ts
import type { EntityRefDto } from './entityRef';
import type { LanguageDto } from './language';

export type Gender = 'Male' | 'Female' | 'Other' | 'Unknown';

export interface CharacterDto {
  id: string;
  name: string;
  worldId: string;
  description?: string;
  firstName?: string;
  lastName?: string;
  middleName?: string;
  maidenName?: string;
  species?: string;
  gender?: Gender;
  birthDate?: string;
  birthDateSuffixId?: string;
  birthDateSuffixAbbreviation?: string;
  deathDate?: string;
  deathDateSuffixId?: string;
  deathDateSuffixAbbreviation?: string;
  birthPlaceLocationId?: string;
  birthPlaceLocationName?: string;
  deathPlaceLocationId?: string;
  deathPlaceLocationName?: string;
  residence?: string;
  biography?: string;
  profilePhotoUrl?: string;
  otherNames?: string;
  position?: string;
  height?: string;
  hairColour?: string;
  parent1Id?: string;
  parent2Id?: string;
  locations: EntityRefDto[];
  affiliations: EntityRefDto[];
  languages: LanguageDto[];
  spouses: EntityRefDto[];
  children: EntityRefDto[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateCharacterDto {
  name: string;
  worldId: string;
  description?: string;
  firstName?: string;
  lastName?: string;
  middleName?: string;
  maidenName?: string;
  species?: string;
  gender?: Gender;
  birthDate?: string;
  birthDateSuffixId?: string;
  deathDate?: string;
  deathDateSuffixId?: string;
  birthPlaceLocationId?: string;
  birthPlaceName?: string;
  deathPlaceLocationId?: string;
  deathPlaceName?: string;
  residence?: string;
  biography?: string;
  otherNames?: string;
  position?: string;
  height?: string;
  hairColour?: string;
  parent1Id?: string;
  parent2Id?: string;
  locationIds?: string[];
  locationNames?: string[];
  affiliationIds?: string[];
  affiliationNames?: string[];
  languageIds?: string[];
  spouseIds?: string[];
  childIds?: string[];
}

export interface UpdateCharacterDto {
  name: string;
  description?: string;
  firstName?: string;
  lastName?: string;
  middleName?: string;
  maidenName?: string;
  species?: string;
  gender?: Gender;
  birthDate?: string;
  birthDateSuffixId?: string | null;
  deathDate?: string;
  deathDateSuffixId?: string | null;
  birthPlaceLocationId?: string | null;
  birthPlaceName?: string;
  deathPlaceLocationId?: string | null;
  deathPlaceName?: string;
  residence?: string;
  biography?: string;
  profilePhotoUrl?: string;
  otherNames?: string;
  position?: string;
  height?: string;
  hairColour?: string;
  parent1Id?: string | null;
  parent2Id?: string | null;
  locationIds: string[];
  locationNames?: string[];
  affiliationIds: string[];
  affiliationNames?: string[];
  languageIds: string[];
  spouseIds: string[];
  childIds: string[];
}